+++ Samples/macOS-Swift/macOS-Swift/CppWrapper.h
@@ -2,4 +2,5 @@
 
 @interface CppWrapper : NSObject
 - (void)throwCPPException;
+- (void)rethrowNoActiveCPPException;
 @end