+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Add start time to network request breadcrumbs (#4008)
+- Add C++ exception support for `__cxa_rethrow` (#3996)
 
 ### Improvements